Seperate loaders into their own file
