+++ app/index.js
@@ -11,6 +11,8 @@ import {Preloader} from './preloader';
 import { SoundBlaster } from './soundblaster';
 import { StageHand } from './stagehand';
 import { Leaf } from './plugin_leaf';
+import { LoadAudio, LoadImages, LoadScript } from './loaders';
+
 
 let imgReady = false;
 let audReady = false;
@@ -34,156 +36,40 @@ let preloader = new Preloader([
 
 let soundBlaster = new SoundBlaster();
 
-///////////////////////////////////////////////////////////////////////////
-// Audio
-///////////////////////////////////////////////////////////////////////////
 
-function LoadAudio(){
 
-    return new Promise(function(resolve, reject){
+function Init(){
 
-        soundBlaster.LoadStream('broadcast',
-            function(event){
-                console.log('load handler called'); 
-                audReady = true;
-                resolve("audio loaded");
-            },
-            function(event){
-                //console.log('time handler called', soundBlaster.GetStreamPosition());
-                if( soundBlaster.GetStreamPosition() > 150 ){
-
+    LoadScript().then(function(script){
+        console.log('LoadScript', script);
+        stageHand = new StageHand(script, plugins);
+        // Display first panel
+        document.querySelector('.panel.one').classList.add('active');
+        // Fetch audio
+        LoadAudio().then(function(){
+            console.log('then after audio ready');
+            StartAudio();
+            tmrIntro = window.setTimeout(function(){
+                if(!audPlaying){
+                    // Show the play button
+                    document.getElementById('manplay').classList.add('active');
                 }
-            },
-            function(event){
-                console.log('play handler called');
-                audPlaying = true;
-                
-                // if( !imgReady ){
-                //     tmrIntro = window.setTimeout(function(){
-                //         // Load the images here
-                //         document.querySelector('.panel.one').classList.remove('active');
-                //         document.querySelector('.panel.one').classList.add('hidden');
-                //         document.querySelector('.panel.two').classList.add('active');
-                //         InitImages();
-                //     },5000);
-                // }
-            },
-            function(event){
-                console.log('ended handler called');
-                // Loop the broadcast
-                soundBlaster.SetStreamPosition(0);
-                soundBlaster.PlayStream();
-                // Reset the stage hand
-                stageHand.Reset();
-            }
-        );
-
+            }, 5000);
+            document.querySelector('.panel.one').classList.remove('active');
+            document.querySelector('.panel.one').classList.add('hidden');
+            document.querySelector('.panel.two').classList.add('active');
+            LoadImages().then(function(){
+                console.log('images ready');
+                BeginProduction();
+            }).catch(function(err){
+                console.error('error loading iamges', err);
+            });
+        });
+    }).catch(function(err){
+        console.error('LoadScript', 'error', err);
     });
 
 }
-
-
-
-///////////////////////////////////////////////////////////////////////////
-// Images
-///////////////////////////////////////////////////////////////////////////
-
-function LoadImages(){
-
-    return new Promise(function(resolve, reject){
-
-        var preloaderPointer = window.setInterval(function(){
-            //console.log('tick', preloader.PercentComplete());
-        },100);
-
-        // tmrIntro = window.setTimeout(function(){
-        //     tmrIntro = null;
-        //     if(imgReady){
-        //         BeginProduction();
-        //     }
-        // }, 5000);
-
-        preloader.PreloadAssets().then(()=>{
-            // TODO: add some sort of check for number of images loaded successfully
-            window.clearInterval(preloaderPointer);
-            preloader.RenderAssets();
-            imgReady = true;
-            resolve(true);
-            // if(tmrIntro === null){
-            //     BeginProduction();
-            // }
-        }).catch((error)=>{
-            reject(error);
-        });
-
-    });
-
-}
-
-///////////////////////////////////////////////////////////////////////////
-// Script
-///////////////////////////////////////////////////////////////////////////
-
-function LoadScript(){
-
-    return new Promise(function(resolve, reject){
-        window.fetch('assets/data/script.json').then(function(response){
-            //console.log('fetch', response);
-            return response.json();
-        })
-        .then(function(data){
-            //console.log('fetch data', data);
-            if( data.acts ){
-                actIdx = data.acts;
-            }else{
-                reject('act data not found in json')
-
-            }
-            if( data.script ){
-                scrReady = true;
-                resolve(data.script);
-            }else{
-                throw('script data not found in json');
-            }
-        })
-        .catch(function(error){
-            reject(error);
-        });
-    });
-
-}
-
-
-LoadScript().then(function(script){
-    console.log('LoadScript', script);
-    stageHand = new StageHand(script, plugins);
-    // Display first panel
-    document.querySelector('.panel.one').classList.add('active');
-    // Fetch audio
-    LoadAudio().then(function(){
-        console.log('then after audio ready');
-        StartAudio();
-        tmrIntro = window.setTimeout(function(){
-            if(!audPlaying){
-                // Show the play button
-                document.getElementById('manplay').classList.add('active');
-            }
-        }, 5000);
-        document.querySelector('.panel.one').classList.remove('active');
-        document.querySelector('.panel.one').classList.add('hidden');
-        document.querySelector('.panel.two').classList.add('active');
-        LoadImages().then(function(){
-            console.log('images ready');
-            BeginProduction();
-        }).catch(function(err){
-            console.error('error loading iamges', err);
-        });
-    });
-}).catch(function(err){
-    console.error('LoadScript', 'error', err);
-});
-
-
 
 function SetScale(){
 	var tscale = 40;
@@ -256,3 +142,4 @@ window.addEventListener('resize', function(event){
 });
 
 SetScale();
+Init();
